+++ src/main/java/com/strangegrotto/wealthdraft/assetallocation/renderer/AssetAllocationRenderer.java
@@ -38,16 +38,18 @@ public class AssetAllocationRenderer {
             var denominatorFilterNameOpt = target.getDenominatorFilterOpt();
             var denominatorStr = denominatorFilterNameOpt.orElse(TOTAL_PORTFOLIO_STR);
 
+            var colorWrapper = new DeviationStatusColorWrapper(calcResult.getDeviationStatus());
+
             var row = ImmAssetAllocationTableRow.of(
-                    numeratorStr,
-                    denominatorStr,
-                    formatCurrencyValue(calcResult.getCurrentNumeratorValue()),
-                    formatCurrencyValue(calcResult.getCurrentDenominatorValue()),
-                    formatFractionAsPercent(calcResult.getCurrentFraction()),
-                    formatFractionAsPercent(calcResult.getTargetFraction()),
-                    formatCurrencyValue(calcResult.getCorrectionNeeded()),
-                    formatFractionAsPercent(calcResult.getDeviationFraction()),
-                    calcResult.getDeviationStatus().toString()
+                    colorWrapper.wrap(numeratorStr),
+                    colorWrapper.wrap(denominatorStr),
+                    colorWrapper.wrap(formatCurrencyValue(calcResult.getCurrentNumeratorValue())),
+                    colorWrapper.wrap(formatCurrencyValue(calcResult.getCurrentDenominatorValue())),
+                    colorWrapper.wrap(formatFractionAsPercent(calcResult.getCurrentFraction())),
+                    colorWrapper.wrap(formatFractionAsPercent(calcResult.getTargetFraction())),
+                    colorWrapper.wrap(formatCurrencyValue(calcResult.getCorrectionNeeded())),
+                    colorWrapper.wrap(formatFractionAsPercent(calcResult.getDeviationFraction())),
+                    colorWrapper.wrap(calcResult.getDeviationStatus().toString())
             );
             table.addRow(row);
         }